pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "video-eval-harness"
version = "0.1.0"
description = "Multi-model video segmentation and labeling benchmark harness"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.25",
    "tenacity>=8.2",
    "pillow>=10.0",
    "opencv-python-headless>=4.8",
    "numpy>=1.24",
    "pandas>=2.0",
    "pyarrow>=14.0",
    "jinja2>=3.1",
    "diskcache>=5.6",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
]
ui = [
    "streamlit>=1.28",
    "plotly>=5.0",
]

[project.scripts]
vbench = "video_eval_harness.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
